% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{ASBVdDetection-package}
\alias{ASBVdDetection}
\alias{ASBVdDetection-package}
\title{ASBVdDetection: X}
\description{
Data
}
\details{
This packages contains functions that I use to push the packages on github.
}
\author{
D. Bonnery
}
\keyword{internal}
