% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_manual.R
\name{build_manual}
\alias{build_manual}
\title{Creates the pdf manual by calling R2pdf}
\usage{
build_manual(path = getwd(), package_name = basename(path))
}
\arguments{
\item{path:}{the root folder of the source package.}

\item{package_name:}{the package name.}
}
\value{
nothing
}
\description{
Creates the pdf manual by calling R2pdf
}
\details{
Creates a pdf file in \code{path}.
}
