% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/butalepaquet.R
\name{butalepaquet}
\alias{butalepaquet}
\title{Buta le paquet sun github}
\usage{
butalepaquet(files.to.exclude = NULL)
}
\arguments{
\item{files.to.exclude}{: a vector of character strings: list of files not to push}
}
\value{
nothing
}
\description{
Buta le paquet sun github
}
\details{
Pushes the package [paquet] on the remote.
this is equivalent to 
git add DESCRIPTION
git add NAMESPACE
git add ./*.pdf"
git add inst/*
git add XXX #(for all files in ./man, ./R, ./demo, ./data that are not listed in files.to.exclude)
git ls-files | grep '\\.Rproj$' | xargs git rm --cached
git rm --cached .gitignore
git rm --cached .Rbuildignore
git rm --cached -r .git
git rm --cached -r .Rproj.user
git add -u
git commit -m 'No information'
git push origin master
}
