% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{EspleitsPelsPaquets-package}
\alias{EspleitsPelsPaquets}
\alias{EspleitsPelsPaquets-package}
\title{EspleitsPelsPaquets: X}
\description{
Data
}
\details{
This packages contains functions that I use to push the packages on github.
}
\author{
D. Bonnery
}
\keyword{internal}
