% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initDESCRIPTION_NAMESPACE.R
\name{initDESCRIPTION_NAMESPACE}
\alias{initDESCRIPTION_NAMESPACE}
\title{Creates DESCRIPTION and NAMESPACE files for a package.}
\usage{
initDESCRIPTION_NAMESPACE(
  packagename,
  destdir = ".",
  packagepath = ".",
  remotes = NULL,
  depends = NULL,
  titlex = "X",
  authorx = "D. Bonnery",
  datex = format(Sys.Date(), "\%Y-\%m-\%d"),
  maintainerx = "D. Bonnery  <daniel.bonnery@gmail.com>",
  description = "Data",
  imports = NULL,
  suggests = NULL,
  LazyLoad = "yes",
  LazyData = "true"
)
}
\arguments{
\item{packagename}{a character string : the package name}

\item{datex}{a character string}

\item{maintainerx}{a character string}

\item{destdir="."}{a character string: the filepath to the destination directory (where the DESCRIPTION and NAMESPACE files will be created)}

\item{packagepath="."}{character string: the filepath to where the package is currently}

\item{remotes=NULL}{a vector of character strings, listing the "remotes" to put in the DESCRIPTION file}

\item{depends=NULL}{a vector of character strings, listing the "depends" to put in the DESCRIPTION file}

\item{titlex="X"}{a character string:}

\item{authorx=}{a character string:}

\item{description=}{a vector of character strings, listing the "depends" to put in the DESCRIPTION file}

\item{imports=NULL, }{a vector of character strings, listing the "depends" to put in the DESCRIPTION file}

\item{suggests=NULL}{a vector of character strings, listing the "depends" to put in the DESCRIPTION file}

\item{LazyLoad="yes"}{a character string, "yes" or "no" listing the "LazyLoad" to put in the DESCRIPTION file}

\item{LazyData="true"}{a character string, "yes" or "no", listing the "LazyData" to put in the DESCRIPTION file}
}
\description{
Creates DESCRIPTION and NAMESPACE files for a package.
}
