% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Initialiselepaquet.R
\name{Initialisalepaquet}
\alias{Initialisalepaquet}
\title{creates the .git initial folder on the machine}
\usage{
Initialisalepaquet(
  paquet,
  remoterepohostingplatform = "github.com",
  account,
  user.name = NULL,
  user.email = NULL
)
}
\arguments{
\item{remoterepohostingplatform}{: character string. Indicates the remote repo hosting platform. "github.com" or "gitlab.com" for example}

\item{account}{: character string the command line will be 'paste0("git remote add origin git@",remoterepohostingplatform,":",account,"/",paquet,".git")}

\item{user.name}{: character string . if not NULL, and init=TRUE, system(paste0('git config  user.name "',user.name,'"')}

\item{user.email}{: character string . if not NULL, and init=TRUE, system(paste0('git config  user.email "',user.email,'"'))}

\item{paquet:}{the package name.}

\item{gitaccount}{: by default paste0(remoterepohostingplatform,":",account)}
}
\value{
nothing
}
\description{
creates the .git initial folder on the machine
}
\details{
Running [Initialisalepaquet(paquet="XXX",
remoterepohostingplatform="github.com",
     account="Tata",
     user.name="Tata Tutu",
     user.email="TataTutu@toto.com")] 
is equivalent of executing in the working directory:
git init 
git confing user.name "Tata Tutu"
git confing user.email "Tata@gmail.com"
git init
git remote add origin git@github.com:Tata/XXX.git
}
