% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tiralepaquet.R
\name{Tiralepaquet}
\alias{Tiralepaquet}
\title{Tira le paquet sun github}
\usage{
Tiralepaquet(
  paquet,
  remoterepohostingplatform = "github.com",
  account,
  gitaccount = paste0(remoterepohostingplatform, ":", account),
  init = FALSE
)
}
\arguments{
\item{remoterepohostingplatform}{: character string. Indicates the remote repo hosting platform. "github.com" or "gitlab.com" for example}

\item{account}{: character string the command line will be 'paste0("git remote add origin git@",remoterepohostingplatform,":",gitaccount,"/",paquet,".git")}

\item{init}{: if TRUE,  git init is executed first}

\item{paquet:}{the package name.}
}
\value{
nothing
}
\description{
Tira le paquet sun github
}
\details{
Pulls the package [paquet] on my github page.
}
